Update and rename security_vulnerability_analysis.py to Automated_vulnerability_scanner_genai.py
This code works if you have weblog data. 
This code finds more than 1,000 consecutive IPs in one second at the same time with each responsesecode value of 404 or 200. 
Next, the generative ai model analyzes the payload of the hacker's vulnerability scan period 
and the payload of the attack after the vulnerability scan. It can even analyze API vulnerability attacks.